복습

+++ 1016_백경현/src/com/bmj/exam/app/Ex13.java
@@ -27,7 +27,7 @@ public class Ex13 {
 				System.out.println( movie.getNo() +")   "+ movie.getName() +
 									" / 감독 : " + movie.getDirector()+
 									" / 개봉일 : " + format.format(movie.getReleaseDate()) +
-									" / 관객수 : " + format1.format((long)movie.getAudienceNum()) +
+									" / 관객수 : " + format1.format(movie.getAudienceNum()) +
 									" / 장르 : " + movie.getGenre());
 			}//for end
 		}
